@@ -1,15 +1,17 @@
 name: bloc_small
 description: An easy-to-use Flutter package offering a streamlined BLoC pattern implementation for intuitive and efficient state management in Flutter apps.
-version: 3.0.1
+version: 3.0.2
 homepage: https://github.com/linhnguyen-gt/bloc_small
 
 environment:
-  sdk: ^3.6.1
+  sdk: ^3.7.2
   flutter: ">=3.3.0"
 
 dependencies:
   flutter:
     sdk: flutter
+  plugin_platform_interface: ^2.0.2
+
   flutter_bloc: ^9.1.0
   get_it: ^8.0.3
   injectable: ^2.5.0
@@ -18,30 +20,14 @@ dependencies:
   bloc_test: ^10.0.0
   freezed_annotation: ^3.0.0
 
-export:
-  flutter_bloc:
-    include: ["lib/bloc.dart"]
-
 dev_dependencies:
   flutter_test:
     sdk: flutter
-  flutter_lints: 5.0.0
+  flutter_lints: ^5.0.0
+
   mockito: ^5.4.5
 # For information on the generic Dart part of this file, see the
 # following page: https://dart.dev/tools/pub/pubspec
-
-# The following section is specific to Flutter packages.
-# flutter:
-# This section identifies this Flutter project as a plugin project.
-# The 'pluginClass' specifies the class (in Java, Kotlin, Swift, Objective-C, etc.)
-# which should be registered in the plugin registry. This is required for
-# using method channels.
-# The Android 'package' specifies package in which the registered class is.
-# This is required for using method channels on Android.
-# The 'ffiPlugin' specifies that native code should be built and bundled.
-# This is required for using `dart:ffi`.
-# All these are used by the tooling to maintain consistency when
-# adding or updating assets for this project.
 
 # To add assets to your plugin package, add an assets section, like this:
 # assets:
