@@ -1,5 +1,4 @@
 import 'package:bloc_small/bloc_small.dart';
-import 'package:injectable/injectable.dart';
 
 part 'search_bloc.freezed.dart';
 part 'search_event.dart';
@@ -16,8 +15,9 @@ class SearchBloc extends MainBloc<SearchEvent, SearchState>
     _searchResults = _searchQuery
         .debounceTime(Duration(milliseconds: 100))
         .doOnData((query) {
-      showLoading(key: 'search');
-    }).switchMap((query) => _performSearch(query));
+          showLoading(key: 'search');
+        })
+        .switchMap((query) => _performSearch(query));
 
     // Listen to search results and update the state
     _searchResults.stream.listen((results) {
@@ -30,14 +30,17 @@ class SearchBloc extends MainBloc<SearchEvent, SearchState>
   }
 
   Future<void> _onUpdateQuery(
-      UpdateQuery event, Emitter<SearchState> emit) async {
+    UpdateQuery event,
+    Emitter<SearchState> emit,
+  ) async {
     await blocCatch(
-        keyLoading: 'search',
-        actions: () async {
-          await Future.delayed(Duration(seconds: 2));
-          _searchQuery.add(event.query);
-        },
-        onError: handleError);
+      keyLoading: 'search',
+      actions: () async {
+        await Future.delayed(Duration(seconds: 2));
+        _searchQuery.add(event.query);
+      },
+      onError: handleError,
+    );
   }
 
   void _onUpdateResults(UpdateResults event, Emitter<SearchState> emit) {
@@ -50,18 +53,21 @@ class SearchBloc extends MainBloc<SearchEvent, SearchState>
 
   ReactiveSubject<List<String>> _performSearch(String query) {
     return ReactiveSubject.fromFutureWithError(
-        Future.delayed(Duration(seconds: 1)).then((_) {
-          if (query.isEmpty) {
-            return <String>[];
-          } else {
-            return ['Result 1 for "$query"', 'Result 2 for "$query"'];
-          }
-        }), onError: (error) {
-      add(SearchError(error.toString()));
-      hideLoading(key: 'search');
-    }, onFinally: () {
-      hideLoading(key: 'search');
-    });
+      Future.delayed(Duration(seconds: 1)).then((_) {
+        if (query.isEmpty) {
+          return <String>[];
+        } else {
+          return ['Result 1 for "$query"', 'Result 2 for "$query"'];
+        }
+      }),
+      onError: (error) {
+        add(SearchError(error.toString()));
+        hideLoading(key: 'search');
+      },
+      onFinally: () {
+        hideLoading(key: 'search');
+      },
+    );
   }
 
   @override
