@@ -1,6 +1,4 @@
-import 'package:auto_route/auto_route.dart';
-import 'package:bloc_small/base/base_app_router.dart';
-import 'package:injectable/injectable.dart';
+import 'package:bloc_small/bloc_small.dart';
 
 import 'app_router.gr.dart';
 
@@ -12,9 +10,9 @@ class AppRouter extends BaseAppRouter {
 
   @override
   List<AutoRoute> get routes => [
-        AutoRoute(page: MyHomeRoute.page, initial: true),
-        AutoRoute(page: SearchRoute.page),
-        AutoRoute(page: ReactiveSubjectScreen.page),
-        AutoRoute(page: CounterRoute.page),
-      ];
+    AutoRoute(page: MyHomeRoute.page, initial: true),
+    AutoRoute(page: SearchRoute.page),
+    AutoRoute(page: ReactiveSubjectScreen.page),
+    AutoRoute(page: CounterRoute.page),
+  ];
 }
