@@ -6,25 +6,30 @@ import 'reactive_subject_drawer.dart';
 class RateLimitedButton extends StatefulWidget {
   static const String route = '/rate_limited_button';
 
+  const RateLimitedButton({super.key});
+
   @override
-  _RateLimitedButtonState createState() => _RateLimitedButtonState();
+  RateLimitedButtonState createState() => RateLimitedButtonState();
 }
 
-class _RateLimitedButtonState extends State<RateLimitedButton> {
+class RateLimitedButtonState extends State<RateLimitedButton> {
   final ReactiveSubject<void> _buttonPressSubject = ReactiveSubject<void>();
   late ReactiveSubject<void> _throttledPressSubject;
-
   @override
   void initState() {
     super.initState();
     // Throttle button presses
-    _throttledPressSubject =
-        _buttonPressSubject.throttleTime(Duration(seconds: 2));
+    _throttledPressSubject = _buttonPressSubject.throttleTime(
+      Duration(seconds: 2),
+    );
     _throttledPressSubject.stream.listen((_) {
       // Handle the button press
-      ScaffoldMessenger.of(context).showSnackBar(
-        SnackBar(content: Text('Button pressed')),
-      );
+      // Add a mounted check before using context
+      if (mounted) {
+        ScaffoldMessenger.of(
+          context,
+        ).showSnackBar(SnackBar(content: Text('Button pressed')));
+      }
     });
   }
 
@@ -43,9 +48,7 @@ class _RateLimitedButtonState extends State<RateLimitedButton> {
   Widget build(BuildContext context) {
     return Scaffold(
       drawer: const ReactiveSubjectDrawer(RateLimitedButton.route),
-      appBar: AppBar(
-        title: Text("Rate Limited Button"),
-      ),
+      appBar: AppBar(title: Text("Rate Limited Button")),
       body: ElevatedButton(
         onPressed: _onButtonPressed,
         child: Text('Press me'),
