@@ -1,9 +1,8 @@
 // example/test/my_bloc_test.dart
 import 'package:bloc_small/bloc_small.dart';
+import 'package:bloc_small_example/bloc/count/count_bloc.dart';
 import 'package:bloc_test/bloc_test.dart';
-import 'package:test/test.dart';
-
-import '../lib/bloc/count/count_bloc.dart';
+import 'package:flutter_test/flutter_test.dart';
 
 void main() {
   group('CountBloc', () {
@@ -21,9 +20,7 @@ void main() {
       build: () => bloc,
       act: (bloc) => bloc.add(Increment()),
       wait: const Duration(seconds: 3),
-      expect: () => [
-        CountState.initial().copyWith(count: 1),
-      ],
+      expect: () => [CountState.initial().copyWith(count: 1)],
     );
 
     tearDown(() {
